+++ backend/app/models/topology.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from app.running_config.util.device_config_types import DeviceConfigInfo
+
+
+class TopologyModel(BaseModel):
+    topology: list[DeviceConfigInfo]
